refactor(training): extrai helpers HTTP para api/_shared.py (Fase 1.2)
Extrai bloco de helpers (~275 linhas) do api/__init__.py para módulo
dedicado api/_shared.py:

- _get_actor_role, _get_actor_id (extração de identidade)
- _session_to_out, _block_to_out, _attendance_to_out,
  _wellness_pre_to_out, _wellness_post_to_out,
  _execution_record_to_out, _session_objective_to_out,
  _mesocycle_to_out, _microcycle_to_out,
  _feedback_thread_to_out, _attention_queue_item_to_out,
  _recommendation_to_out, _ineligibility_to_out
- _feedback_context_type, _feedback_context_ref_id

__init__.py: 1844 → 1586 linhas.

Invariantes preservadas:
- test_route_inventory_frozen: PASS
- test_public_surface: PASS
- Suite training/tests: 251 passed

Próximo: Fase 1.3 (decomposição por subdomínio).
